refactor: `ERROR_MESSAGE` 상수 import문 변경
