+++ src/utils/constants/index.ts
@@ -22,3 +22,4 @@ export {
   ZINDEX_MARKER_MOUSE_OUT,
 } from '@/utils/constants/map';
 export { BACKGROUND_ANIMATION, SCROLL_STYLE } from '@/utils/constants/style';
+export { ERROR_MESSAGE } from '@/utils/constants/message';
